Obj::Type not actually used - stripping it out Splitting upvalue to its own file
